Import Socket io and api url to MainContent
deleteRoom lives inside main MainContent
